move pull v2 code in a separate file
